requirements.txt
langchain
uvicorn==0.24.0
python-dotenv==1.0.0
# google-generativeai==0.3.1
# tiktoken==0.5.2
matplotlib
plotly
openai==0.27.8
pandas
numpy
scipy
scikit-learn
# flask
# flask_cors
fastapi
